+++ src/core/mod.rs
@@ -10,6 +10,7 @@ pub mod fri;
 pub mod oods;
 pub mod poly;
 pub mod proof_of_work;
+pub mod prover;
 pub mod queries;
 pub mod utils;
